pyproject.toml
[project]
name = "dotflow"
version = "0.15.0.dev6"
authors = [
  { name="Fernando Celmer", email="email@fernandocelmer.com" },
]
license = {file = "LICENSE"}
description = "🎲 Dotflow turns an idea into flow! Lightweight Python library for execution pipelines with retry, parallel, cron and async support."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
	"rich",
  "pydantic",
  "typing-extensions",
  "cookiecutter (>=2.0)",
  "requests"
]
keywords = [
  "pipeline",
  "workflow",
  "etl",
  "task-runner",
  "orchestration",
  "automation",
  "parallel",
  "async",
  "retry",
  "data-pipeline",
  "scheduler",
  "cron",
  "cron-job",
]
classifiers = [
  'Development Status :: 4 - Beta',
  "Operating System :: OS Independent",
  'Intended Audience :: Developers',
  'Natural Language :: English',
  "License :: OSI Approved :: MIT License",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: System :: Distributed Computing",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/dotflow-io/dotflow"
Issues = "https://github.com/dotflow-io/dotflow/issues"
Repository = "https://github.com/dotflow-io/dotflow"
Documentation = "https://dotflow-io.github.io/dotflow/"
Changelog = "https://dotflow-io.github.io/dotflow/nav/development/release-notes/"

[project.optional-dependencies]
aws = ["boto3"]
gcp = ["google-cloud-storage"]
deploy-aws = ["boto3"]
deploy-gcp = ["google-cloud-run", "google-cloud-build", "google-cloud-artifact-registry", "google-cloud-service-usage", "google-cloud-storage", "google-auth"]
deploy-alibaba = ["alibabacloud-fc-open20210406", "alibabacloud-tea-openapi"]
deploy-github = ["PyGithub", "GitPython"]
scheduler = ["croniter"]
otel = ["opentelemetry-api", "opentelemetry-sdk"]
sentry = ["sentry-sdk"]

[tool.poetry]
name = "dotflow"
version = "0.15.0.dev6"
description = "🎲 Dotflow turns an idea into flow! Lightweight Python library for execution pipelines with retry, parallel, cron and async support."
authors = ["Fernando Celmer <email@fernandocelmer.com>"]
readme = "README.md"
license = "MIT"
package-mode = true
homepage = "https://github.com/dotflow-io/dotflow"
documentation = "https://dotflow-io.github.io/dotflow/"
repository = "https://github.com/dotflow-io/dotflow"
keywords = [
  "pipeline",
  "workflow",
  "etl",
  "task-runner",
  "orchestration",
  "automation",
  "parallel",
  "async",
  "retry",
  "data-pipeline",
  "scheduler",
  "cron",
  "cron-job",
]
classifiers = [
  'Development Status :: 4 - Beta',
  "Operating System :: OS Independent",
  'Intended Audience :: Developers',
  'Natural Language :: English',
  "License :: OSI Approved :: MIT License",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: System :: Distributed Computing",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]

[tool.poetry.scripts]
dotflow = 'dotflow.main:main'

[tool.poetry.dependencies]
python = ">=3.9.0"
rich = "^14.1.0"
pydantic = "^2.11.7"
typing-extensions = "^4.14.1"
python-dotenv = "^1.1.0"
requests = "^2.32.4"

[tool.poetry.group.dev.dependencies]
build = "^1.2.2.post1"
pytest = "^8.3.4"
flake8 = "^7.1.1"
tox = "^4.26.0"
pytest-cov = "^6.0.0"
pylint = "^3.3.4"
pyzmq = "^26.2.1"
boto3 = "^1.42.83"
moto = "^5.1.22"


[tool.poetry.group.code-quality.dependencies]
ruff = "^0.9.0"
mypy = "^1.13.0"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocstrings = {extras = ["python"], version = "^0.29.0"}
griffe-typingdoc = "^0.2.8"
mike = "^2.1.3"
mkdocs-simple-blog = "0.2.0"
mkdocs-material = "^9.6.13"
markdown-include-variants = "^0.0.4"
mkdocs-static-i18n = "^1.3.0"
mkdocs-redirects = "^1.2.2"
mdx-include = "^1.4.2"
griffe = "<1.14"


[[tool.poetry.source]]
name = "test"
url = "https://test.pypi.org/simple/"
priority = "supplemental"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
